# backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FoodFlow API",
    "version": "1.0.0",
    "description": "API para o sistema FoodFlow, um SaaS para gestão de restaurantes. Inclui autenticação, gestão de restaurantes, usuários, mesas, cardápio, sessões, pedidos, relatórios, reservas e personalização de estilo."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    },
    {
      "url": "https://foodflow-api.render.com",
      "description": "Servidor de produção (Render)"
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica um usuário",
        "description": "Autentica um usuário com e-mail e senha, retornando um token JWT para acesso aos endpoints protegidos.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              },
              "example": {
                "email": "dono@restaurante.com",
                "password": "senha123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login bem-sucedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                },
                "example": {
                  "message": "Login bem-sucedido",
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Credenciais inválidas"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Cadastra um novo usuário",
        "description": "Cria um novo usuário (admin, waiter, cook, customer). O restaurant_id é opcional para admin. Use /roles para obter os IDs dos papéis.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInput"
              },
              "example": {
                "email": "dono@restaurante.com",
                "password": "senha123",
                "name": "João Silva",
                "role_id": "09603787-2fca-4e4c-9e6c-7b349232c512",
                "restaurant_id": null
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                },
                "example": {
                  "id": "user_1",
                  "email": "dono@restaurante.com",
                  "name": "João Silva",
                  "role_id": "09603787-2fca-4e4c-9e6c-7b349232c512",
                  "restaurant_id": null
                }
              }
            }
          },
          "400": {
            "description": "Erro no cadastro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "errors": [
                    {
                      "msg": "role_id inválido",
                      "path": "role_id",
                      "location": "body"
                    },
                    {
                      "msg": "restaurant_id inválido",
                      "path": "restaurant_id",
                      "location": "body"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Envia e-mail para redefinição de senha",
        "description": "Envia um e-mail com um link para redefinição de senha para o endereço fornecido.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "E-mail do usuário"
                  }
                },
                "required": [
                  "email"
                ]
              },
              "example": {
                "email": "dono@restaurante.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "E-mail enviado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "message": "E-mail de redefinição enviado"
                }
              }
            }
          },
          "400": {
            "description": "Erro ao enviar e-mail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "E-mail não encontrado"
                }
              }
            }
          }
        }
      }
    },
    "/auth/profile": {
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Atualiza o perfil do usuário",
        "description": "Atualiza o nome ou e-mail do usuário autenticado. Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileInput"
              },
              "example": {
                "name": "João Silva",
                "email": "novo.dono@restaurante.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                },
                "example": {
                  "id": "user_1",
                  "email": "novo.dono@restaurante.com",
                  "name": "João Silva",
                  "role_id": "09603787-2fca-4e4c-9e6c-7b349232c512",
                  "restaurant_id": null
                }
              }
            }
          },
          "400": {
            "description": "Erro ao atualizar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "E-mail já em uso"
                }
              }
            }
          }
        }
      }
    },
...
    "/customization": {
      "post": {
        "tags": [
          "Customization"
        ],
        "summary": "Cria ou atualiza configurações de estilo",
        "description": "Define ou atualiza as configurações de estilo (cores, logo) de um restaurante. Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomizationInput"
              },
              "example": {
                "restaurant_id": "rest_1",
                "primary_color": "#FF5733",
                "secondary_color": "#C70039",
                "logo_url": "https://github.com/seu-usuario/foodflow-backend/logo.png"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Configuração criada ou atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomizationResponse"
                },
                "example": {
                  "restaurant_id": "rest_1",
                  "primary_color": "#FF5733",
                  "secondary_color": "#C70039",
                  "logo_url": "https://github.com/seu-usuario/foodflow-backend/logo.png"
                }
              }
            }
          },
          "400": {
            "description": "Erro ao criar/atualizar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Restaurante não encontrado"
                }
              }
            }
          }
        }
      }
    },
    "/customization/{restaurantId}": {
      "get": {
        "tags": [
          "Customization"
        ],
        "summary": "Obtém configurações de estilo",
        "description": "Retorna as configurações de estilo de um restaurante (público, usado por clientes).",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID do restaurante"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configurações de estilo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomizationResponse"
                },
                "example": {
                  "restaurant_id": "rest_1",
                  "primary_color": "#FF5733",
                  "secondary_color": "#C70039",
                  "logo_url": "https://github.com/seu-usuario/foodflow-backend/logo.png"
                }
              }
            }
          },
          "404": {
            "description": "Configuração não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Configuração não encontrada"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Customization"
        ],
        "summary": "Exclui configurações de estilo",
        "description": "Remove as configurações de estilo de um restaurante. Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID do restaurante"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Configuração excluída"
          },
          "400": {
            "description": "Erro ao excluir",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Configuração não encontrada"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerauth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via /auth/login"
      }
    },
    "schemas": {
...
      "LoginInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "E-mail do usuário"
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Mensagem de sucesso"
          },
          "token": {
            "type": "string",
            "description": "Token JWT para autenticação"
          }
        },
        "required": [
          "message",
          "token"
        ]
      },
      "RegisterInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "E-mail do usuário"
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário"
          },
          "name": {
            "type": "string",
            "description": "Nome do usuário"
          },
          "role_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do papel"
          },
          "restaurant_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do restaurante (opcional para admin)",
            "nullable": true
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "role_id"
        ]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do usuário"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "E-mail do usuário"
          },
          "name": {
            "type": "string",
            "description": "Nome do usuário"
          },
          "role_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do papel"
          },
          "restaurant_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do restaurante",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role_id"
        ]
      },
      "UpdateProfileInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do usuário",
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "E-mail do usuário",
            "nullable": true
          }
        }
      },
...
      "CustomizationInput": {
        "type": "object",
        "properties": {
          "restaurant_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do restaurante"
          },
          "primary_color": {
            "type": "string",
            "description": "Cor primária (hexadecimal)"
          },
          "secondary_color": {
            "type": "string",
            "description": "Cor secundária (hexadecimal)"
          },
          "logo_url": {
            "type": "string",
            "format": "uri",
            "description": "URL do logo",
            "nullable": true
          }
        },
        "required": [
          "restaurant_id",
          "primary_color",
          "secondary_color"
        ]
      },
      "CustomizationResponse": {
        "type": "object",
        "properties": {
          "restaurant_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do restaurante"
          },
          "primary_color": {
            "type": "string",
            "description": "Cor primária (hexadecimal)"
          },
          "secondary_color": {
            "type": "string",
            "description": "Cor secundária (hexadecimal)"
          },
          "logo_url": {
            "type": "string",
            "format": "uri",
            "description": "URL do logo",
            "nullable": true
          }
        },
        "required": [
          "restaurant_id",
          "primary_color",
          "secondary_color"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mensagem de erro",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "description": "Mensagem de erro específica"
                },
                "path": {
                  "type": "string",
                  "description": "Campo com erro"
                },
                "location": {
                  "type": "string",
                  "description": "Local do erro (ex.: body, query)"
                }
              }
            },
            "description": "Lista de erros detalhados",
            "nullable": true
          }
        }
      }
    }
  }
}